+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * distribute_commons_on_links: only put boardings on first links and alightings on last link (when there is multiple link merge together to create a common_link, not between trips) 
 * preparation_footpaths: if length=0: dont do any foothpaths
 * create_common_links:fix bug where time was already in the common_trips. so it was not recomputed on aggregation.
+* uses zone_to_road walk_time if avaible in pt_pathfinder and analysis_pt_time
 
 ## Changes
 * road_pathfinder: allow zone-to-zone in the expanded graph.